third_party/blink/web_tests/virtual/without-coep-for-shared-worker/external/wpt/html/cross-origin-embedder-policy/shared-workers.https-expected.txt

This is a testharness.js-based test.
PASS default policy (derived from response)
FAIL "require-corp" (derived from response) assert_equals: expected "failure" but got "success"
PASS default policy (derived from owner set due to use of local scheme - blob URL)
FAIL "require-corp" (derived from owner set due to use of local scheme - blob URL) assert_equals: expected "failure" but got "success"
PASS default policy (derived from owner set due to use of local scheme - data URL)
FAIL "require-corp" (derived from owner set due to use of local scheme - data URL) assert_equals: expected "failure" but got "success"
Harness: the test ran to completion.
